requirements.txt
streamlit==1.32.0
torch==1.13.1
torchvision==0.14.1
pillow==10.1.0
pandas==2.1.1
plotly==5.18.0
scikit-learn==1.3.0
numpy==1.26.0